Use size_t to be consistent (#775)
#716

Created by bonsai from df8abad7f4d3dbf9d08b3f3c3723f9bcf78cba0c

+++ sql/item_func.cc
@@ -4859,7 +4859,7 @@ void udf_handler::set_vdf_string_result_type(Item *item,
   // set_data_type_string then buckets to VARCHAR / MEDIUM / LONG blob by
   // length.
   // TODO(villagesql): Allow VDFs to choose an encoding.
-  uint64_t len = 0;
+  size_t len = 0;
   if (u_d != nullptr && u_d->vdf_func_desc != nullptr &&
       u_d->vdf_func_desc->protocol >= VEF_PROTOCOL_4) {
     len = u_d->vdf_func_desc->max_result_length;
